Multiple changes
Fixed when zip_buses.sh attempts to read collected.txt or
not_collected.txt and it doesn't exists
Fixed main config path on PopulateDatabase.py

+++ app/zip_buses.sh
@@ -7,8 +7,26 @@ OLD_FOLDER=$FOLDER'_old'
 YESTERDAY=$(date --date="1 day ago" +"%Y-%m-%d")
 tar -zcvf $OLD_FOLDER'/'$YESTERDAY".tar.gz" $FOLDER'/'$YESTERDAY
 rm -r $FOLDER'/'$YESTERDAY
-echo "collected $(wc -l $COLLECTED) entries" && rm $COLLECTED && touch $COLLECTED
-echo "failed $(wc -l $NOT_COLLECTED) entries" && rm $NOT_COLLECTED && touch $NOT_COLLECTED
-echo "$YESTERDAY,$(wc -l $COLLECTED),$(wc -l $NOT_COLLECTED)" >> $OLD_FOLDER'/collection_logs.csv'
+
+
+RESULTS="$YESTERDAY,"
+if [[ -e $COLLECTED ]]; then
+	RESULTS=$RESULTS$(wc -l $COLLECTED)","
+	echo "collected $(wc -l $COLLECTED) entries" && rm $COLLECTED && touch $COLLECTED
+else
+	RESULTS=$RESULTS"0,"
+	touch $COLLECTED
+fi
+
+if [[ -e $NOT_COLLECTED ]]; then
+	RESULTS=$RESULTS$(wc -l $NOT_COLLECTED)
+	echo "failed $(wc -l $NOT_COLLECTED) entries" && rm $NOT_COLLECTED && touch $NOT_COLLECTED
+else
+	RESULTS=$RESULTS"0"
+	touch $NOT_COLLECTED
+fi
+
+echo $RESULTS >> $OLD_FOLDER'/collection_logs.csv'
+
 # Execute script per-day automaticaly
 python3 /app/PopulateDatabase.py -b $OLD_FOLDER -d $YESTERDAY bus
